+++ app/Models/Package.php
@@ -2,11 +2,18 @@
 
 namespace App\Models;
 
+use App\Traits\LogsActivity;
 use Illuminate\Database\Eloquent\Model;
 use Illuminate\Database\Eloquent\Relations\HasMany;
 
 class Package extends Model
 {
+    use LogsActivity;
+
+    public function activityDescription(): string
+    {
+        return $this->name;
+    }
     protected $fillable = [
         'name',
         'description',